Add weeks, months, and years, ago, from now, before, ahead
about a week ago
3 weeks from now
about a month before
3 months ahead
about a year
about a year and 3 months
3 years and about a month

There's a gap for 1yr and 1mo. Will return
"about a year and about a month ago"
and "3 years and about a month" should be formated
as "about 3 years and a month".  But you're welcome
to fork and add support and issue a pull request.