notifier.fire(*e*); and shrinkage.
No need to maintain separate detach handle
properties on unique subs, so moved _attach()
to on(), aliased delegate() to on() and
detachDelegate() to detach().

Also now using the triggering key event as
the DOMEventFacade seed in notifier.fire(e);
This removed the facade nesting for callbacks
whereby e._event was pointing at a generated
event facade instead of the native DOM event.