Add gallery patch to create o.td in 3.4.1
3.5.0 will see more significant changes to DT,
so rather than apply a fix in master to the
missing/wrong o.td bug that would encourange
poor performance practices and is equally likely
to go away in 3.5.0, 3.4.1 will see o.td removed.

Formatters will not have a broken property to use,
and must instead use this gallery module to allow
them to create a Node for each cell.

In reality, a Node is still created for each cell,
so performance will continue to be poor, but the
less implementers rely on that infrastructure, the
better off they will be when string concat rendering
comes along in 3.5.0.