Change of heart; abort.
The gallery module had only two methods,
comprised of a few lines of code each, and
it made sense to include the logic from one
method in datatable-base, leaving only one
method with one line of code.  The method
didn't seem so bad to include in base, and
the story for DT 3.4.1- is certainly simpler.

The usual caveats about DT changing in 3.5.0
still apply, so I'm not too worried about it.